app/services/voice_service.py
import os
from groq import Groq

def transcribe_audio(file_path: str) -> str:
    """
    Transcribes audio file to text using Groq's Whisper model.
    """
    api_key = os.getenv("GROQ_API_KEY")
    if not api_key:
        raise ValueError("GROQ_API_KEY not found in environment.")

    client = Groq(api_key=api_key)
    
    with open(file_path, "rb") as file:
        transcription = client.audio.transcriptions.create(
            file=(os.path.basename(file_path), file.read()),
            model="whisper-large-v3",
            response_format="text"
        )
    
    return transcription
